REAL quantity type

+++ twoface/db/quantity_type.py
@@ -1,5 +1,5 @@
 # Third-party
-from sqlalchemy.types import TypeDecorator, Numeric
+from sqlalchemy.types import TypeDecorator, REAL
 from astropy.units import Quantity
 
 __all__ = ['QuantityTypeClassFactory']
@@ -8,7 +8,7 @@ def QuantityTypeClassFactory(unit):
     class QuantityType(TypeDecorator):
         """ Custom type to handle `~astropy.units.Quantity` objects. """
 
-        impl = Numeric
+        impl = REAL
 
         def process_bind_param(self, value, dialect):
             if isinstance(value, Quantity):
